pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "marimo[recommended]>=0.20.4",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "plotly>=6.6.0",
    "pmtiles>=3.7.0",
    "polars>=1.38.1",
    "rasterio>=1.5.0",
    "requests>=2.32.0",
]
